update_stock_data/requirements.txt
yfinance
google-cloud-bigquery
numpy
